upgrade to pybind11, and use its setuptools utilities
