+++ pyproject.toml
@@ -1,0 +1,9 @@
+[build-system]
+requires = [
+    "setuptools>=42",
+    "wheel",
+    "pybind11>=2.6.1",
+    "numpy",
+]
+
+build-backend = "setuptools.build_meta"
